transfering everything into an iframe based site
i got the header looking just about right, still need to load the font for the pause/play button, but overall very responsive, very happy with it.
although using an iframe will severly limit compatibility with old ass browsers and people who have js disabled, so thats a bit of a bummer.
i still need to work on the music player, i think i want to base it off of similar code from octogonis 2, where theres a json file on the server that the
music player grabs from and uses that info to load in new tracks and display them when on the music page, possibly a bit overkill but its fun. maybe i can add cookie support so it can save your position when you get back! that would be neat